refactor: clean up RedisOMProperties
